==> 01-proce-data-conclat.qmd <==
# Procesamiento Data Conclat

```{r,echo=FALSE,warning=FALSE,message=FALSE}
rm(list=ls())
options(scipen = 999)

library(pacman)
p_load(bookdown,
       car,
       flextable,
       ggalluvial,
       ggplot2,
       ggrepel,
       gridExtra,
       kableExtra,
       knitr,
       labelled,
       lubridate,
       officer,
       psych,
       readr,
       readxl,
       scales,         # para etiquetas en gráficos
       shadowtext,
       sjmisc,
       sjPlot,
       sjlabelled,
       statar,
       stringr,
       survey,        # diseño survey
       srvyr,         # diseño survey con tidy
       tidyverse,
       tinytex,
       viridis,
       httr
       )
```

## Cargar datos

```{r datos,message=FALSE,warning=FALSE}
load("input/data/datos-conclat.Rdata")
data_raw <- data
```

## Selección de variables

```{r data,message=FALSE,warning=FALSE}
data <- data %>% 
  select(id,
         rake_wb2,
         # Relación con los medios de producción
         a07,                         # sit empleo
         b19,                         # sindicatos en empresa 
         b21,                         # afiliación sindical
         # Políticas
         c01_01,                      # Importancia: mejores salarios
         c01_02,                      # Importancia: reducir dedsigualdad
         c05_02,                      # Acuerdo: desigualda es porque dueños empresa pagan poco
         c05_03,                      # Acuerdo: alto nivel vida dueños empresas es por pagar poco
         c06_02,                      # Acuerdo: apoyo sistema solidario de reparto
         c07_01,                      # Acuerdo: aumentar impuestos a los más ricos
         c07_03,                      # Acuerdo: sindicatos tengan más poder
         c07_04,                      # Acuerdo: redistribuir riquezas
         c07_05,                      # Acuerdo: igualdad hombres y mujeres
         c07_06,                      # Acuerdo: educación universitaria pública
         c08,                         # posición política    
         c10_01,                      # habla de politica con amigos
         c15_01,                      # confia en partidos políticos
         # Demográficas
         sexo=e01,                    # sexo
         edad=e02,                    # edad
         e05,                         # educación  
         a25,a25_esp,a25_esp_rec,
         a26,
         a27)# %>% 
  #sjlabelled::set_na(.,na=c(-999,-888,-777))
```

## Procesamiento

### sit_empleo

```{r,warning=FALSE,message=FALSE}
# Situación empleo
data <- data %>%
  mutate(
    a07 = factor(
      as.numeric(a07),
      levels = 1:9,
      labels = c(
        "1. Patrón(a) o empleador(a)",
        "2. Trabajador(a) por cuenta propia",
        "3. Empleado(a) u obrero(a) del sector público (Gobierno Central o Municipal)",
        "4. Empleado(a) u obrero(a) de empresas públicas",
        "5. Empleado(a) u obrero(a) del sector privado",
        "6. Servicio doméstico puertas adentro",
        "7. Servicio doméstico puertas afuera",
        "8. FF.AA. y del Orden",
        "9. Familiar no remunerado"
      )
    )
  ) %>%
  set_variable_labels(a07 = "Employment situation (a07)")

frq(data$a07)

data <- data %>%
  mutate(
    sit_empleo = case_when(
      as_numeric(a07) == 1 ~ 1,
      as_numeric(a07) == 2 ~ 2,
      as_numeric(a07) >= 3 ~ 3,
      TRUE ~ NA_real_
      ),
    sit_empleo = factor(
      sit_empleo,
      levels = 1:3,
      labels = c(
        "Empleador(a)",
        "Autoempleado(a)",
        "Empleado(a)"
      )
    )
  ) %>%
  set_variable_labels(sit_empleo = "Relación con los medios de producción")

frq(data$sit_empleo)
```

### Filtrar por sit_empleo

```{r}
# data <- data %>% 
#   filter(sit_empleo == "Empleado(a)")
```

### sindicato

```{r,warning=FALSE,message=FALSE}
frq(data$b19)

# Sindicatos en empresas
data <- data %>%
  mutate(
    sindicato = case_when(
      as.numeric(sit_empleo) %in% c(1, 2) ~ NA_real_,
      as.numeric(sit_empleo) == 3 & b19 == 1 ~ 1,
      as.numeric(sit_empleo) == 3 & b19 %in% c(2, -888, -999) ~ 0,
      TRUE ~ NA_real_
    )
  ) %>%
  set_variable_labels(sindicato = "Presencia de sindicato en empresa")

frq(data$sindicato)
```

### af_sindical

```{r,warning=FALSE,message=FALSE}
frq(data$b21)

# Afiliación sindical
data <- data %>%
  mutate(
    af_sindical = case_when(
      as.numeric(sit_empleo) %in% c(1, 2) ~ NA_real_,
      as.numeric(sit_empleo) == 3 & b21 %in% c(1, 2) ~ 1,
      as.numeric(sit_empleo) == 3 & b21 %in% c(3, -888, -999) ~ 0,
      sindicato == 0 ~ 0
    )) %>%
  set_variable_labels(af_sindical = "Afiliación sindical")

frq(data$af_sindical)
```

### sexo

```{r}
frq(data$sexo)

data <- data %>%
  mutate(
    sexo = factor(
      sexo, 
      levels = 1:2,
      labels = c(
        "Hombre",
        "Mujer"
      )
      )
    ) %>%
  set_variable_labels(sexo = "Sexo")

frq(data$sexo)
```

### edad

```{r}
descr(data$edad)

data <- data %>%
  mutate(
    edad = ifelse(edad %in% c(-888, -999), NA, edad),
    edad_tramos = cut(
      edad,
      breaks = c(17, 24, 34, 44, 54, Inf),
      labels = c("18-24", "25-34", "35-44", "45-54", "55 o +"),
      right = TRUE,
      include.lowest = TRUE
    )
  ) %>%
  set_variable_labels(edad_tramos = "Tramos de edad")

frq(data$edad_tramos)
```

### c01_01

```{r}
# c01_01: Importancia: mejores salarios
frq(data$c01_01)

data <- data %>%
  mutate(
    c01_01 = ifelse(c01_01 %in% c(-999, -888), NA, c01_01)
  ) %>%
  set_variable_labels(c01_01 = "Nivel de importancia: Salarios y condiciones laborales (c01_01)")

frq(data$c01_01)
```

### c01_02

```{r}
frq(data$c01_02)

data <- data %>%
  mutate(
    c01_02 = ifelse(c01_02 %in% c(-999, -888), NA, c01_02)
  ) %>%
  set_variable_labels(c01_02 = "Nivel de mportancia: Reducir desigualdad (c01_02)")

frq(data$c01_02)
```

### c05_02

```{r}
# c05_02: Acuerdo: desigualdad es porque dueños empresa pagan poco
frq(data$c05_02)

data <- data %>%
  mutate(
    c05_02 = ifelse(c05_02 %in% c(-999, -888), NA, c05_02)
  ) %>%
  set_variable_labels(c05_02 = "Nivel de acuerdo: desigualdad es porque dueños empresa pagan poco (c05_02)")

frq(data$c05_02)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c05_02_acuerdo = ifelse(
      c05_02 %in% c(4, 5), 1,
      ifelse(c05_02 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c05_02_acuerdo = "De acuerdo o Muy de acuerdo: desigualdad es porque dueños empresa pagan poco")

frq(data$c05_02_acuerdo)
```

### c05_03

```{r}
# c05:03: Acuerdo: alto nivel vida dueños empresas es por pagar poco
frq(data$c05_03)

data <- data %>%
  mutate(
    c05_03 = ifelse(c05_03 %in% c(-999, -888), NA, c05_03)
  ) %>%
  set_variable_labels(c05_03 = "Nivel de acuerdo: alto nivel vida dueños empresas es por pagar poco (c05_03)")

frq(data$c05_03)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c05_03_acuerdo = ifelse(
      c05_03 %in% c(4, 5), 1,
      ifelse(c05_03 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c05_03_acuerdo = "De acuerdo o Muy de acuerdo: alto nivel vida dueños empresas es por pagar poco")

frq(data$c05_03_acuerdo)
```

### c06_02

```{r}
# c06_02: Acuerdo: apoyo sistema solidario de reparto
frq(data$c06_02)

data <- data %>%
  mutate(
    c06_02 = ifelse(c06_02 %in% c(-999, -888), NA, c06_02)
  ) %>%
  set_variable_labels(c06_02 = "Nivel de acuerdo: apoyo sistema solidario de reparto (c06_02)")

frq(data$c06_02)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c06_02_acuerdo = ifelse(
      c06_02 %in% c(4, 5), 1,
      ifelse(c06_02 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c06_02_acuerdo = "De acuerdo o Muy de acuerdo: apoyo sistema solidario de reparto")

frq(data$c06_02_acuerdo)
```

### c07_01

```{r}
# Acuerdo: aumentar impuestos a los más ricos
frq(data$c07_01)

data <- data %>%
  mutate(
    c07_01 = ifelse(c07_01 %in% c(-999, -888), NA, c07_01)
  ) %>%
  set_variable_labels(c07_01 = "Nivel de acuerdo: aumentar impuestos a los más ricos (c07_01)")

frq(data$c07_01)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c07_01_acuerdo = ifelse(
      c07_01 %in% c(4, 5), 1,
      ifelse(c07_01 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c07_01_acuerdo = "De acuerdo o Muy de acuerdo: aumentar impuestos a los más ricos")

frq(data$c07_01_acuerdo)
```

### c07_03

```{r}
# Acuerdo: sindicatos tengan más poder
frq(data$c07_03)

data <- data %>%
  mutate(
    c07_03 = ifelse(c07_03 %in% c(-999, -888), NA, c07_03)
  ) %>%
  set_variable_labels(c07_03 = "Nivel de acuerdo: sindicatos tengan más poder(c07_03)")

frq(data$c07_03)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c07_03_acuerdo = ifelse(
      c07_03 %in% c(4, 5), 1,
      ifelse(c07_03 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c07_03_acuerdo = "De acuerdo o Muy de acuerdo: sindicatos tengan más poder")

frq(data$c07_03_acuerdo)
```

### c07_04

```{r}
# Acuerdo: redistribuir riquezas
frq(data$c07_04)

data <- data %>%
  mutate(
    c07_04 = ifelse(c07_04 %in% c(-999, -888), NA, c07_04)
  ) %>%
  set_variable_labels(c07_04 = "Nivel de acuerdo: redistribuir riquezas (c07_04)")

frq(data$c07_04)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c07_04_acuerdo = ifelse(
      c07_04 %in% c(4, 5), 1,
      ifelse(c07_04 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c07_04_acuerdo = "De acuerdo o Muy de acuerdo: redistribuir riquezas")

frq(data$c07_04_acuerdo)
```

### c07_05

```{r}
# Acuerdo: igualdad hombres y mujeres
frq(data$c07_05)

data <- data %>%
  mutate(
    c07_05 = ifelse(c07_05 %in% c(-999, -888), NA, c07_05)
  ) %>%
  set_variable_labels(c07_05 = "Nivel de acuerdo: igualdad hombres y mujeres (c07_05)")

frq(data$c07_05)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c07_05_acuerdo = ifelse(
      c07_05 %in% c(4, 5), 1,
      ifelse(c07_05 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c07_05_acuerdo = "De acuerdo o Muy de acuerdo: igualdad hombres y mujeres")

frq(data$c07_05_acuerdo)
```

### c07_06

```{r}
# Acuerdo: educación universitaria pública
frq(data$c07_06)

data <- data %>%
  mutate(
    c07_06 = ifelse(c07_06 %in% c(-999, -888), NA, c07_06)
  ) %>%
  set_variable_labels(c07_06 = "Nivel de acuerdo: educación universitaria pública (c07_06)")

frq(data$c07_06)

# Crear variable dicotómica: acuerdo (4 o 5)
data <- data %>%
  mutate(
    c07_06_acuerdo = ifelse(
      c07_06 %in% c(4, 5), 1,
      ifelse(c07_06 %in% c(1, 2, 3), 0, NA))
  )%>%
  set_variable_labels(c07_06_acuerdo = "De acuerdo o Muy de acuerdo: educación universitaria pública")

frq(data$c07_06_acuerdo)
```

### pol_id

```{r}
frq(data$c08)

data <- data %>%
  mutate(
    pol_id = case_when(
      c08 %in% c(0, 1, 2, 3) ~ 1,
      c08 %in% c(4, 5, 6) ~ 2,
      c08 %in% c(7, 8, 9, 10) ~ 3,
      c08 %in% c(11, 12, -888, -999) ~ 4,
      TRUE ~ NA_real_
    ),
    pol_id = factor(
      pol_id, 
      levels = 1:4, 
      labels = c(
        "Izquierda", 
        "Centro", 
        "Derecha", 
        "No se identifica"
        ) 
      )
    ) %>%
  set_variable_labels(pol_id = "Identificación política") 

frq(data$pol_id)
```

### politizado

```{r}
frq(data$c10_01)
frq(data$c15_01)

data <- data %>%
  mutate(
    politizado = if_else(
      c10_01 %in% c(4, 5) & c15_01 %in% c(3, 4, 5),1,0
    )
  ) %>% 
  set_variable_labels(politizado = "Politizado o No politizado") %>% 
  set_value_labels(politizado = c(
    "Sí" = 1,
    "No" = 0
  ))

frq(data$politizado)
```

### clases

```{r}
# Cargar data clases
data_clases <- readRDS("input/data/datos-clases.rds")

# Seleccionar solo clave y variables
datos_clases_sel <- data_clases %>%
  select(id, wrightclass1_reduced, wrightclass2_reduced)

# Unir a base 
data <- data %>%
  left_join(datos_clases_sel, by = "id")
```

## Guardar datos

```{r}
save(data,file="output/data/datos-cut.Rdata")
saveRDS(data,file="output/data/datos-cut.rds")
```
